Move Pointer, PointerBlock to separate module.
